+++ r/NEWS.md
@@ -19,10 +19,12 @@
 
 # arrow 7.0.0.9000
 
+* `read_csv_arrow()`'s readr-style type `T` is now mapped to `timestamp(unit = "ns")` instead of `timestamp(unit = "s")`.
 * `lubridate`: 
   * `tz()` to extract/get timezone
   * `semester()` to extract/get semester
   * `dst()` to get daylight savings time indicator.
+  * `epiyear()` to get epiyear
 
 # arrow 7.0.0